requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0.6
pycodestyle==2.11.1
pytz==2023.3
sqlparse==0.4.4
Unipath==1.1
dj-database-url==2.1.0
python-decouple==3.8
gunicorn==21.2.0
Pillow==10.3.0
whitenoise==6.5.0
# psycopg2-binary==2.9.7  # Uncomment for PostgreSQL production deployment on Render
sentence-transformers>=2.2.0  # Optional: For AI-powered grant analysis (downloads ~100MB model on first use)
